+++ hutool-crypto/src/main/java/cn/hutool/crypto/asymmetric/AsymmetricAlgorithm.java
@@ -12,6 +12,8 @@ public enum AsymmetricAlgorithm {
 	RSA("RSA"), 
 	/** RSA算法，此算法用了默认补位方式为RSA/ECB/PKCS1Padding */
 	RSA_ECB_PKCS1("RSA/ECB/PKCS1Padding"), 
+	/** RSA算法，此算法用了默认补位方式为RSA/ECB/NoPadding */
+	RSA_ECB("RSA/ECB/NoPadding"),
 	/** RSA算法，此算法用了RSA/None/NoPadding */
 	RSA_None("RSA/None/NoPadding");