+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 新特性
 * 【extra  】     增加自动装配SpringUtil类（pr#1366@Github）
 * 【extra  】     ArrayUtil增加map方法重载
+* 【crypto 】     AsymmetricAlgorithm增加RSA_ECB("RSA/ECB/NoPadding")（issue#1368@Github）
 
 ### Bug修复
 * 【core   】     修复FileUtil.move以及PathUtil.copy等无法自动创建父目录的问题（issue#I2CKTI@Gitee）